requirements.txt
# ====================================================================================================
# Project Requirements — GenericBoilerplatePython
# ----------------------------------------------------------------------------------------------------
# Install all dependencies:
#     pip install -r requirements.txt
#
# Notes:
# - snowflake-connector-python[secure-local-storage] enables secure Okta token caching.
# - google-api-* packages support Drive/Sheets/Forms integration.
# - tkcalendar powers date selection widgets in the GUI.
# - tqdm, requests, openpyxl, and PyPDF2 are used broadly across multiple modules.
# - selenium and webdriver-manager support browser automation and scraping workflows.
# - This file intentionally omits built-in modules (e.g. tkinter, zipfile, tempfile).
# ====================================================================================================

# ----------------------------------------------------------------------------------------------------
# 1. CORE DATA & UTILITIES
# ----------------------------------------------------------------------------------------------------
pandas>=2.3
numpy>=1.26
tqdm>=4.66
requests>=2.32
pyyaml>=6.0.2

# ----------------------------------------------------------------------------------------------------
# 2. FILE & DOCUMENT HANDLING
# ----------------------------------------------------------------------------------------------------
pdfplumber>=0.11
PyPDF2>=3.0
pdfminer.six>=20231228       # Installed automatically with pdfplumber, pinned for stability
openpyxl>=3.1

# ----------------------------------------------------------------------------------------------------
# 3. DATABASE & CONNECTORS
# ----------------------------------------------------------------------------------------------------
snowflake-connector-python[secure-local-storage]>=4.0

# ----------------------------------------------------------------------------------------------------
# 4. GUI & FRONTEND COMPONENTS
# ----------------------------------------------------------------------------------------------------
tkcalendar>=1.6.1

# ----------------------------------------------------------------------------------------------------
# 5. WEB & AUTOMATION
# ----------------------------------------------------------------------------------------------------
selenium>=4.25               # Browser automation (Chrome, Edge, Firefox)
webdriver-manager>=4.0       # Automatically manages browser drivers

# ----------------------------------------------------------------------------------------------------
# 6. GOOGLE API & AUTH
# ----------------------------------------------------------------------------------------------------
google-api-python-client>=2.100
google-auth>=2.40
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2

# ----------------------------------------------------------------------------------------------------
# 7. PACKAGING, BUILD & DEV TOOLS
# ----------------------------------------------------------------------------------------------------
pyinstaller>=6.0             # For building standalone executables